chore(server): move middleware to middleware/
- auth
- routeLogger
