+++ server/routes/api/email.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 
-const auth = require('../auth')
+const auth = require('../../middleware/auth')
 const emailController = require('../controllers/emailController')._init()
 
 router.route('/')
