+++ server/routes/api/projects.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 
-const auth = require('../auth')
+const auth = require('../../middleware/auth')
 const projectsController = require('../controllers/projectsController')._init()
 
 router.route('/')
