+++ server/routes/api/users.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
 
-const auth = require('../auth')
+const auth = require('../../middleware/auth')
 const usersController = require('../controllers/usersController')._init()
 
 // router.param('user', usersController.preloadUser)
